Plan 9 from Bell Labs Fourth Edition 2005-10-28

+++ sys/src/libc/alpha/_seek.c
@@ -1,7 +1,7 @@
 #include <u.h>
 #include <libc.h>
 
-extern	long	_seek(vlong*, int, vlong, int);
+extern	int	_seek(vlong*, int, vlong, int);
 
 vlong
 seek(int fd, vlong o, int p)
@@ -9,6 +9,6 @@ seek(int fd, vlong o, int p)
 	vlong l;
 
 	if(_seek(&l, fd, o, p) < 0)
-		l = -1;
+		l = -1LL;
 	return l;
 }
